Fix user self-deleting and changing ADMIN role

+++ src/main/java/com/online/wallet/service/UserService.java
@@ -78,9 +78,11 @@ public class UserService implements UserDetailsService {
         .orElseThrow(() -> new UserNotFoundException("User with id: " + id + " not found"));
   }
 
-  public void deleteUserById(Long userID) {
-    findUserById(userID);
-    userRepository.deleteById(userID);
+  public void deleteUserById(User user, Long userID) {
+      if (user != null && !user.getId().equals(userID)) {
+          findUserById(userID);
+          userRepository.deleteById(userID);
+      }
   }
 
   public List<User> findAllUsers() {
